refactor(types, packages): reflect commented feedbacks
