+++ apps/web/layouts/template-create/BlockGroupModifier.tsx
@@ -8,31 +8,15 @@ import { BlockImageModifier } from './BlockImageModifier';
 import { DefaultBlockModifier } from './DefaultBlockModifier';
 import { DefaultGroupModifier } from './DefaultGroupModifier';
 import { TextBlockModifier } from './TextBlockModifier';
-import { BlockSubType } from './types';
-
-export interface BlockModifierFactoryPropsInterface {
-  subType: BlockSubType;
-}
-
-export interface BlockModifierPropsInterface extends BlockModifierFactoryPropsInterface {
-  type: 'block';
-}
-
-export interface GroupModifierFactoryPropsInterface {
-  subType?: 'animator';
-}
-
-export interface GroupModifierPropsInterface extends GroupModifierFactoryPropsInterface {
-  type: 'group';
-}
-
-export type BlockGroupModifierPropsInterface =
-  | BlockModifierPropsInterface
-  | GroupModifierPropsInterface;
+import {
+  BlockGroupModifierPropsInterface,
+  BlockModifierFactoryPropsInterface,
+  GroupModifierFactoryPropsInterface,
+} from './types';
 
 /**
  * @description
- * 아직 그룹에 대한 명세가 많이 없지만,s
+ * 아직 그룹에 대한 명세가 많이 없지만,
  * 추후 확장 가능성이 있어 미리 수정하기 용이하도록 만듣다.
  */
 function GroupModifierFactory({ subType }: GroupModifierFactoryPropsInterface) {
