+++ apps/web/templates/layouts/sidebar/right/index.ts
@@ -3,3 +3,4 @@ export * from './ActivedGroupColorsModifier';
 export * from './ActivedBlockFillModifier';
 export * from './ActivedBlockBorderModifier';
 export * from './ActivedTextModifier';
+export * from './ActivedImageModifier';
